### 2024-01-25 #6283 @editorialbot
[PRE REVIEW]: TDApplied: An R package for machine learning and inference with persistence diagrams
**Submitting author:** @shaelebrown (<a href="http://orcid.org/0000-0001-8868-2867">Shael Brown</a>)
**Repository:** https://github.com/shaelebrown/TDApplied
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v3.0.2
