### 2024-01-25 #6283 @danielskatz
👋 @shaelebrown - thanks for your submission. You'll need to license it with an OSI-approved license before we can proceed.  I also wonder if you might want to add any graphics to your paper to make the work more clear to those outside the field?
