### 2024-01-25 #6283 @danielskatz
👋 @shaelebrown - in addition, `@editorialbot` checks all references in the references file. Since many of yours are not actually cited in the paper, you might want to make a paper.bib file that just contains the ones you need, and use that in the paper.

Either way, you will have to fix some DOIs, as shown in the result of the `@editorialbot check references` command, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right (or if you change the .md file, eg. by adding a figure) to make a new PDF. editorialbot commands need to be the first entry in a new comment.

When you are ready to proceed (no correct missing or invalid DOIs, an open source license, perhaps a figure in the paper), please ping me.
