### 2024-01-26 #6283 @shaelebrown
Hi Daniel,

Thanks for that! I thought that GPL3 would be covered but I can see that it
is not. Do I need to change the license on just my github repo or submit a
package update to CRAN with the new license? Looking to add a graphic now..

On Thu, Jan 25, 2024 at 3:00 PM Daniel S. Katz ***@***.***>
wrote:

> 👋 @shaelebrown <https://github.com/shaelebrown> - thanks for your
> submission. You'll need to license it with an OSI-approved license before
> we can proceed. I also wonder if you might want to add any graphics to your
> paper to make the work more clear to those outside the field?
>
