### 2024-01-26 #6283 @danielskatz
using GPL3 is fine, but I didn't see it in the repo. Now I see it in the DESCRIPTION file.  You could also add a license file (see https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository) to make GitHub recognize the license, but it's not technically needed for JOSS.
