+++ DESCRIPTION
@@ -15,7 +15,7 @@ Description: Topological data analysis is a powerful tool for finding non-linear
     topological data analysis pipelines.
 Depends: R (>= 3.5.0)
 Imports: parallel, doParallel, foreach, clue, rdist, parallelly, kernlab, iterators, methods, stats, utils, Rcpp (>= 0.11.0)
-License: GPL-3
+License: GPL (>= 3)
 URL: https://github.com/shaelebrown/TDApplied
 BugReports: https://github.com/shaelebrown/TDApplied/issues
 Encoding: UTF-8
