### 2024-01-28 8a2ce5c
Updated software bibliography and added a paper figure.
