### 2024-01-28 5b2d8fb
Fixed figure caption
