### 2024-01-28 #6283 @shaelebrown
Thanks Daniel. I've added a license file to the JOSS branch (not the main
branch) to ease with the review, fixed and checked the bibliography and
added a figure. Please let me know if all looks good and we can proceed,
thanks!

On Fri, Jan 26, 2024 at 11:01 AM Daniel S. Katz ***@***.***>
wrote:

> using GPL3 is fine, but I didn't see it in the repo. Now I see it in the
> DESCRIPTION file. You could also add a license file (see
> https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/adding-a-license-to-a-repository)
> to make GitHub recognize the license, but it's not technically needed for
> JOSS.
>
