### 2024-01-28 #6283 @danielskatz
👋 @mahfuz05062 - As a new JOSS editor, would this submission be of interest to you as your first?
