### 2024-01-28 #6283 @mahfuz05062
@danielskatz Sure. I don't have experience with persistence diagrams, but I have good experience with R and some experience with MDS so I should be good.
