### 2024-01-28 #6283 @shaelebrown
Wonderful! @mahfuz05062 I look forward to hearing your input. Daniel I can
see that @EduPH has expertise in R and topological data analysis (the other
members of the list with similar credentials I have collaborated with,
rrrlw and corybrunson). Some other options would be the contributor of the
R package TDA Jisu Kim (https://jkim82133.github.io/research.html), the
creator of the Ball mapper algorithm and R package Pawel Dlotko (
https://dioscuri-tda.org/members/pawel.html), an author of several TDA R
software packages Matt Pieckenbrock (https://peekxc.github.io), but I'm
open to others. If @lmcinnes or @sauln are open to coming to the dark side
of R then they would also be good options.

On Sun, Jan 28, 2024 at 12:42 PM Daniel S. Katz ***@***.***>
wrote:

> @shaelebrown <https://github.com/shaelebrown> - thanks, this all looks
> good. I'll proceed to looking for an editor. While I do so, if you have any
> suggestions for potential reviewers then please mention them here in this
> thread (without tagging them with an @). You can search the list of people
> <https://reviewers.joss.theoj.org/lookup> that have already agreed to
> review and may be suitable for this submission, or make suggestions from
> outside this list.
>
