### 2024-01-28 #6283 @danielskatz
@shaelebrown - just a reminder, please don't tag potential reviewers - we don't need to add notifications to anyone until the editor decides on who to invite...

> @shaelebrown - thanks, this all looks good. I'll proceed to looking for an editor. While I do so, if you have any suggestions for potential reviewers then please mention them here in this thread (**without tagging them with an @**). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission, or make suggestions from outside this list.
