### 2024-01-28 #6283 @shaelebrown
So sorry about that, will make sure to avoid this going forward.

On Sun, Jan 28, 2024 at 2:40 PM Daniel S. Katz ***@***.***>
wrote:

> @shaelebrown <https://github.com/shaelebrown> - just a reminder, please
> don't tag potential reviewers - we don't need to add notifications to
> anyone until the editor decides on who to invite...
>
> @shaelebrown <https://github.com/shaelebrown> - thanks, this all looks
> good. I'll proceed to looking for an editor. While I do so, if you have any
> suggestions for potential reviewers then please mention them here in this
> thread (*without tagging them with an @*). You can search the list of
> people <https://reviewers.joss.theoj.org/lookup> that have already agreed
> to review and may be suitable for this submission, or make suggestions from
> outside this list.
>
