### 2024-02-04 #6283 @peekxc
> Hi @peekxc, welcome to JOSS! Thank you for agreeing to review this submission over email. Please confirm again here for visibility.

Confirmed
