### 2024-02-05 #6321 @mahfuz05062
As mentioned above, you can use the command `@editorialbot generate my checklist` to create your review checklist.  As you go over the submission, please check any items that you feel have been satisfied.

There are also links to the JOSS reviewer guidelines (https://joss.readthedocs.io/en/latest/reviewer_guidelines.html)

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#6321` so that a link is created to this thread for visibility. Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about **2-4 weeks**. Please let me know if you require additional time. We can also use `editorialbot` (our bot) to set automatic reminders if you know you'll be away for a known period.
