### 2024-02-22 shaelebrown/TDApplied#1 @peekxc
LICENSE file 
Hyy @shaelebrown,

Could you add plain-text LICENSE or COPYING file to the repo, or are you fine explicitly leaving one out?

JOSS requires such a file, and for repos hosted on GH it's nice to have, as the license is detected and shown immediately in a 'tab' next to the README.

CRAN standards allow omitting a License file if its generic enough, but you can also point to the LICENSE file directly via 'file LICENSE' (see https://cran.r-project.org/doc/manuals/r-devel/R-exts.html#Licensing-1)

See also [#6321](https://github.com/openjournals/joss-reviews/issues/6321#issuecomment-1938909073)
