### 2024-02-22 shaelebrown/tdaml#1 @shaelebrown
Hey @peekxc

I didn't originally have a license file but I did add one at the start of the review process: https://github.com/shaelebrown/TDApplied/blob/JOSS/LICENSE.md

Let me know if that's alright!
