### 2024-02-24 shaelebrown/TDApplied#2 @peekxc
How to contribute? 
@shaelebrown Could you add a note to the README.md about how to contribute, or consider adding a [contributing.md](https://mozillascience.github.io/working-open-workshop/contributing/)? I apologize if this is shown elsewhere

[#6321](https://github.com/openjournals/joss-reviews/issues/6321)
