### 2024-02-24 shaelebrown/TDApplied#3 @peekxc
Documentation outline
While each of the individual functions in TDApplied are well documented (e.g. analyze_representatives, plot_diagrams, etc.), it would be nice to have an outline view of all the functions somewhere, e.g. in the package overview ("TDApplied::`TDApplied-package`").

Similarly, while there _are_ links to other package functions in the descriptions of the arguments and details, I think the designated ways of doing this that are a bit more common (see: https://cran.r-project.org/web/packages/roxygen2/vignettes/index-crossref.html)

[#6321](https://github.com/openjournals/joss-reviews/issues/6321)
