### 2024-02-25 shaelebrown/tdaml#2 @shaelebrown
Great point @peekxc. Do you know if I update the main branch or the JOSS branch?
