### 2024-02-26 shaelebrown/tdaml#2 @shaelebrown
The JOSS branch is the most up to date for this review, but I'm not sure if whatever changes I make should be on the main branch for a future CRAN update or not
