### 2024-02-26 shaelebrown/tdaml#2 @peekxc
Isn't the goal to eventually merge the JOSS branch into main?

I don't think it matters which one you modify. Once they merge, it'll make no difference

( I might be misunderstanding something here )
