### 2024-02-29 #6321 @mahfuz05062
> I guess I can add it here. Related to the software paper references. It would be nice to add TDA manuals such as (at least one):
>
> * Computational Topology - an Introduction. H. Edelsbrunner, and J. Harer. American Mathematical Society, (2010)
> * Dey TK, Wang Y. Computational Topology for Data Analysis. Cambridge University Press; 2022.
> * Carlsson G, Vejdemo-Johansson M. Topological Data Analysis with Applications. Cambridge University Press; 2021.

@shaelebrown can you address this? Thanks!
