### 2024-03-04 shaelebrown/tdaml#2 @shaelebrown
I think you're right. I've updated the README.Rmd and README.md files with notes on how to contribute. I'll close the issue for now but let me know if I should add in anything else!
