### 2024-03-04 f6011cd
Added TDA manual citation
