+++ vignettes/REFERENCES.bib
@@ -118,6 +118,15 @@ title = {Means and medians of sets of persistence diagrams}
   url = {https://gudhi.inria.fr/}
 }
 
+@book{TDA, 
+place={Cambridge}, 
+title={Topological Data Analysis with Applications}, 
+DOI={10.1017/9781108975704}, 
+publisher={Cambridge University Press}, 
+author={Carlsson, Gunnar and Vejdemo-Johansson, Mikael}, 
+year={2021}
+}
+
 @Manual{PHAT,
   title = {Persistent homology algorithm toolbox},
   author = {Ulrich Bauer and Michael Kerber and Jan Reininghaus},
