### 2024-03-04 shaelebrown/tdaml#3 @shaelebrown
That's a great comment, thanks. Can you send me an example from another package? I've never seen that before.

In terms of the cross references I followed the conventions in this link https://cran.r-project.org/doc/manuals/R-exts.html#Cross_002dreferences to make xrefs. When I tried to follow the instructions in your link you sent, updating the analyze_representatives.R file, running `devtools::document()` and check the `help("analyze_representatives")` help page it doesn't seem to format properly (see screen grab):

What do you suggest?
