### 2024-03-04 shaelebrown/tdaml#3 @peekxc
> Can you send me an example from another package? I've never seen that before.

Sure, you can try ?ggplot2::`ggplot2-package` to get an idea of simple one. To see how to set it up, see e.g. [here](https://r-pkgs.org/man.html#sec-man-package-doc).

A more informal example that doesn't generate an entire package-specific page is `?dplyr::select`'s page. There's not only a nice overview of the functions it composes with (contains(), matches(), etc.), but also a nice "see also" section linking to other single-verb functions that behave similarly (arrange(), mutate(), filter(), etc.). 'See also' links are mentioned in the same CRAN xref page you linked.

Similar functionality in TDApplied might benefit from such "see also" linking; `diagram_mds`, `diagram_kpca` are both projections of diagrams, `permutation_test` and `independence_test` are both types of hypothesis tests for diagrams, etc.

> When I tried to follow the instructions in your link you sent, updating the analyze_representatives.R file, running devtools::document() and check the help("analyze_representatives") help page it doesn't seem to format properly (see screen grab):

I guess the exact syntax may differ based on the roxygen version or something, whatever way of making cross references works for you!
