### 2024-03-06 7a34381
Added @seealso documentation and package description function.
