+++ R/TDApplied-package.R
@@ -1,0 +1,8 @@
+#' @useDynLib TDApplied, .registration = TRUE
+#' @docType package
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL
