### 2024-03-06 shaelebrown/tdaml#3 @shaelebrown
Awesome. Added in the `TDApplied-package` function, and added `seealso` linking in function documentation. The cross refs are working for me so I left them unchanged. Let me know if all looks good and I'll close the issue!
