+++ vignettes/paper.bib
@@ -1,147 +1,0 @@
-@Manual{R-TDA,
-  title = {TDA: Statistical Tools for Topological Data Analysis},
-  author = {Brittany T. Fasy and Jisu Kim and Fabrizio Lecci and Clement Maria and David L. Millman and Vincent Rouvreau.},
-  year = {2021},
-  note = {R package version 1.7.7},
-  url = {https://CRAN.R-project.org/package=TDA},
-}
-
-@Manual{R-TDAstats,
-  title = {TDAstats: Pipeline for Topological Data Analysis},
-  author = {Raoul Wadhwa and Andrew Dhawan and Drew Williamson and Jacob Scott},
-  year = {2019},
-  note = {R package version 0.4.1},
-  url = {https://github.com/rrrlw/TDAstats},
-}
-
-@Article{TDAstats2018,
-  title = {TDAstats: R pipeline for computing persistent homology in topological data analysis},
-  author = {Raoul R. Wadhwa and Drew F. K. Williamson and Andrew Dhawan and Jacob G. Scott},
-  journal = {Journal of Open Source Software},
-  year = {2018},
-  volume = {3},
-  number = {28},
-  pages = {860},
-  url = {https://doi.org/10.21105/joss.00860},
-  doi = {10.21105/joss.00860},
-}
-
-@article{PHoriginal,
-  title={Topological Persistence and Simplification},
-  author={Herbert Edelsbrunner and David Letscher and Afra Zomorodian},
-  journal={Discrete \& Computational Geometry},
-  year={2000},
-  volume={28},
-  pages={511-533},
-  doi={10.1007/s00454-002-2885-2}
-}
-
-@article{ComputingPH,
-author = {Zomorodian, Afra and Carlsson, Gunnar},
-year = {2005},
-month = {02},
-pages = {249-274},
-title = {Computing Persistent Homology},
-volume = {33},
-journal = {Discrete and Computational Geometry},
-doi = {10.1007/s00454-004-1146-y}
-}
-
-@Manual{R-devtools,
-    title = {devtools: Tools to Make Developing R Packages Easier},
-    author = {Hadley Wickham and Jim Hester and Winston Chang and Jennifer Bryan},
-    year = {2021},
-    note = {R package version 2.4.3},
-    url = {https://CRAN.R-project.org/package=devtools},
-  }
-
-@Article{Robinson_Turner,
-author = {Robinson, Andrew and Turner, Katharine},
-year = {2017},
-title = {Hypothesis testing for topological data analysis},
-journal = {Journal of Applied and Computational Topology},
-volume = {1},
-issue = {2}
-}
-
-@inproceedings{persistence_fisher,
- author = {Le, Tam and Yamada, Makoto},
- booktitle = {Advances in Neural Information Processing Systems},
- editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
- pages = {},
- publisher = {Curran Associates, Inc.},
- title = {Persistence Fisher Kernel: A Riemannian Manifold Kernel for Persistence Diagrams},
- url = {https://proceedings.neurips.cc/paper/2018/file/959ab9a0695c467e7caf75431a872e5c-Paper.pdf},
- volume = {31},
- year = {2018}
-}
-
-@article {TDA_ADHD,
-author = {Gracia-Tabuenca, Zeus and Diaz-Patino, Juan Carlos and Arelio, Isaac and Alcauter, Sarael},
-title = {Topological Data Analysis reveals robust alterations in the whole-brain and frontal lobe functional connectomes in Attention-Deficit/Hyperactivity Disorder},
-elocation-id = {751008},
[…]
