### 2024-03-13 #6321 @shaelebrown
@mahfuz05062 I confirm that:

1. The authors and ORCID ID's are correct.
2. TDApplied version 3.0.3 has been published to CRAN with the updates made in this review process.
3. The zenodo DOI is 10.5281/zenodo.10814141 (https://doi.org/10.5281/zenodo.10814141).
4. The authors and ORCID ID's from the release are correct.
5. The license in the release is correct.

I'll just need to update the paper.md file with those kinds of information, or does that get handled automatically?
