### 2024-03-25 7c0ec3a
Adding DOIs to paper refs
