+++ vignettes/paper.bib
@@ -61,7 +61,8 @@ year = {2017},
 title = {Hypothesis testing for topological data analysis},
 journal = {Journal of Applied and Computational Topology},
 volume = {1},
-issue = {2}
+issue = {2},
+doi = {10.1007/s41468-017-0008-7}
 }
 
 @inproceedings{persistence_fisher,
@@ -73,7 +74,8 @@ issue = {2}
  title = {Persistence Fisher Kernel: A Riemannian Manifold Kernel for Persistence Diagrams},
  url = {https://proceedings.neurips.cc/paper/2018/file/959ab9a0695c467e7caf75431a872e5c-Paper.pdf},
  volume = {31},
- year = {2018}
+ year = {2018},
+ doi = {10.48550/arXiv.1802.03569}
 }
 
 @article {TDA_ADHD,
