+++ vignettes/paper.bib
@@ -48,7 +48,7 @@ doi = {10.1007/s00454-004-1146-y}
 }
 
 @Manual{R-devtools,
-    title = {devtools: Tools to Make Developing R Packages Easier},
+    title = {{d}evtools: Tools to Make Developing {R} Packages Easier},
     author = {Hadley Wickham and Jim Hester and Winston Chang and Jennifer Bryan},
     year = {2021},
     note = {R package version 2.4.3},
@@ -71,7 +71,7 @@ doi = {10.1007/s41468-017-0008-7}
  editor = {S. Bengio and H. Wallach and H. Larochelle and K. Grauman and N. Cesa-Bianchi and R. Garnett},
  pages = {},
  publisher = {Curran Associates, Inc.},
- title = {Persistence Fisher Kernel: A Riemannian Manifold Kernel for Persistence Diagrams},
+ title = {Persistence {F}isher Kernel: A {R}iemannian Manifold Kernel for Persistence Diagrams},
  url = {https://proceedings.neurips.cc/paper/2018/file/959ab9a0695c467e7caf75431a872e5c-Paper.pdf},
  volume = {31},
  year = {2018},
@@ -89,7 +89,7 @@ doi = {10.1523/eneuro.0543-19.2020}
 
 @Article{R-testthat,
     author = {Hadley Wickham},
-    title = {testthat: Get Started with Testing},
+    title = {{t}estthat: Get Started with Testing},
     journal = {The R Journal},
     year = {2011},
     volume = {3},
