### 2024-03-26 #6321 @danielskatz
@shaelebrown - I've also suggested some small changes in https://github.com/shaelebrown/TDApplied/pull/5 - please merge this, or let me know what you disagree with, then we can proceed.
