### 2024-03-28 #6321 @shaelebrown
@danielskatz @mahfuz05062 @peekxc @EduPH thank you all so much for your invaluable feedback, suggestions, time and effort. It is a pleasure to see our paper up on JOSS! Thank you all!
